requirements.txt
boto3
datasets
Flask
openai
pandas
pillow
PyMuPDF
tqdm
numpy
requests
scikit-learn
tqdm
gunicorn
pyalex
flask-cors
pymongo
pyjwt
langchain_anthropic
langchain_openai
langchain_text_splitters
pypdf
pytz
pydantic
pymongo
langchain_core
langchain_community
langchain
pyalex
passlib
stripe
rapidfuzz
azure-storage-blob
azure-search-documents
transformers
torch
langchain_google_vertexai
vertexai
